requirements.txt
pandas
sqlalchemy
mysql-connector-python
reportlab
textblob
python-dotenv 